Fixed bug #66891 (Unexpected HTTP 401 after NTLM authentication)
If you set a "WWW-Authenticate" header PHP automatically issues a 401.

http\Env\Response slurps any set headers and re-sends it after sending
the status code, so that PHP might be resetting the status code to 401
because it sees again an WWW-Authenticate header.

Send the status code after any headers to avoid this issue.